third_party/WebKit/LayoutTests/imported/wpt/service-workers/service-worker/postmessage-to-client.https-expected.txt

This is a testharness.js-based test.
FAIL postMessage from ServiceWorker to Client assert_equals: message event's origin should be set correctly. expected "https://127.0.0.1:8444" but got "https://web-platform.test:8444"
Harness: the test ran to completion.
